refactor: home quantize, batch, boot, and serve tests
Move the remaining large inline test modules next to their impls:
quantize/src/tests.rs, batch/tests.rs, boot/tests.rs, serve/tests.rs.
